requirements.txt
streamlit
groq
pdfplumber
python-dotenv
requests
fpdf2
streamlit-javascript
